+++ src/types.ts
@@ -1,6 +1,12 @@
 import Grid from "@egjs/grid";
+import { InfiniteGridItem } from "./InfiniteGridItem";
 
 export interface InfiniteGridGroup {
   groupKey: string | number;
   grid: Grid;
+}
+
+export interface CategorizedGroup {
+  groupKey: number | string;
+  items: InfiniteGridItem[];
 }
